+++ src/rewards/components/attributes/AddUpcast.tsx
@@ -1,8 +1,13 @@
 import React from "react";
 import AddRemoveButton from "../../../components/AddRemoveButton";
 import AttributeDescription from "../AttributeDescription";
-import { ChangeValueFunc, RewardData } from "../../types/reward-types";
+import {
+  ChangeValueFunc,
+  REWARD_TYPE,
+  RewardData,
+} from "../../types/reward-types";
 import styles from "./../AddAttributes.module.scss";
+import { initReward } from "../../util/reward-calcs";
 
 interface AddUpcastAttributeProps {
   selectedOptions: RewardData;
@@ -20,6 +25,20 @@ const AddUpcastAttribute: React.FC<AddUpcastAttributeProps> = ({
   setUpcastReward,
 }) => {
   if (selectedOptions.upcast) return null;
+  const upcastRewards = rewards
+    .map((rewardData, i) => {
+      return {
+        rewardIndex: i,
+        rewardData,
+      };
+    })
+    .filter(({ rewardData }) => {
+      const reward = initReward(rewardData);
+      return reward.tier === -1 && reward.type === REWARD_TYPE.FEATURE;
+    })
+    .sort((a, b) =>
+      (a.rewardData.name || "").localeCompare(b.rewardData.name || "")
+    );
 
   return (
     <li>
@@ -33,9 +52,9 @@ const AddUpcastAttribute: React.FC<AddUpcastAttributeProps> = ({
         onChange={(e) => setUpcastReward(parseInt(e.target.value))}
       >
         <option value={-1}>Select a reward</option>
-        {rewards.map((reward, i) => (
-          <option key={i} value={i}>
-            {reward.name}
+        {upcastRewards.map(({ rewardData, rewardIndex }) => (
+          <option key={rewardIndex} value={rewardIndex}>
+            {rewardData.name}
           </option>
         ))}
       </select>