+++ src/CHANGELOG.md
@@ -6,11 +6,14 @@
 - Change stunned to have a tier limit and can be applied multiple times
 - Decrease stunned from +3 to +2 tiers
 - Change restrained and stunned breaking conditions
+- Add Chewie rules
+- Change range increase to +2
 
 ## 0.3.2
 
 - Add half tiers
 - Set enemy wellspring to tier x 2
+- Change difficulty DCs to 10 + tier
 
 ## 0.3.1